SYAL-1232 [PROD-PWC] Symphony generates unnecessary tickets when Cisco WebEx Room Kits are in presentation mode.
