+++ src/main/java/com/avispl/dal/communicator/cisco/dto/status/teams/MicrosoftTeamsHardwareAcceleratorDecoder.java
@@ -1,11 +1,11 @@
 /*
  * Copyright (c) 2024 AVI-SPL Inc. All Rights Reserved.
  */
-package com.avispl.dal.communicator.cisco.dto.status.webrtc.teams;
+package com.avispl.dal.communicator.cisco.dto.status.teams;
 
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty;
 
-public class MicrosoftTeamsHardwareAcceleratorDecoder extends MicrosoftTeamsHardwareAcceleratorMetrics{
+public class MicrosoftTeamsHardwareAcceleratorDecoder extends MicrosoftTeamsHardwareAcceleratorMetrics {
     @JacksonXmlProperty(localName = "CroppedHeight")
     private String croppedHeight;
     @JacksonXmlProperty(localName = "CroppedWidth")
