+++ src/main/java/com/avispl/dal/communicator/cisco/dto/status/teams/MicrosoftTeamsUser.java
@@ -1,7 +1,7 @@
 /*
  * Copyright (c) 2024 AVI-SPL Inc. All Rights Reserved.
  */
-package com.avispl.dal.communicator.cisco.dto.status.webrtc.teams;
+package com.avispl.dal.communicator.cisco.dto.status.teams;
 
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty;
